EPMRPP-97027 implement GET user avatar endpoint
