+++ src/main/java/com/epam/ta/reportportal/core/configs/MvcConfig.java
@@ -47,6 +47,7 @@ import org.springframework.core.annotation.Order;
 import org.springframework.http.MediaType;
 import org.springframework.http.converter.ByteArrayHttpMessageConverter;
 import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.ResourceHttpMessageConverter;
 import org.springframework.http.converter.StringHttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.validation.beanvalidation.BeanValidationPostProcessor;
@@ -121,6 +122,7 @@ public class MvcConfig implements WebMvcConfigurer {
   @Override
   public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
     converters.clear();
+    converters.add(resourceConverter());
     converters.add(byteArrayConverter());
     converters.add(jsonConverter());
     converters.add(openMetricsTextStringConverter());
@@ -171,6 +173,11 @@ public class MvcConfig implements WebMvcConfigurer {
     StringHttpMessageConverter converter = new StringHttpMessageConverter();
     converter.setSupportedMediaTypes(Collections.singletonList(MediaType.TEXT_PLAIN));
     return converter;
+  }
+
+  @Bean
+  public ResourceHttpMessageConverter resourceConverter() {
+    return new ResourceHttpMessageConverter();
   }
 
   @Bean
