+++ src/main/java/com/epam/ta/reportportal/core/file/GetFileHandler.java
@@ -35,6 +35,15 @@ public interface GetFileHandler {
   BinaryData getUserPhoto(ReportPortalUser loggedInUser, boolean loadThumbnail);
 
   /**
+   * Returns {@link InputStream} for current logged-in user photo
+   *
+   * @param userId  requested user Id {@link Long}
+   * @param loadThumbnail true if needed to load thumbnail
+   * @return {@link BinaryData}
+   */
+  BinaryData getUserPhoto(Long userId, boolean loadThumbnail);
+
+  /**
    * Returns {@link InputStream} for photo of the {@link com.epam.ta.reportportal.entity.user.User}
    * with specified username
    *
