+++ src/test/java/com/epam/ta/reportportal/ws/controller/UserControllerTest.java
@@ -36,12 +36,7 @@ import com.epam.reportportal.model.ValidationConstraints;
 import com.epam.ta.reportportal.dao.IssueTypeRepository;
 import com.epam.ta.reportportal.dao.ProjectRepository;
 import com.epam.ta.reportportal.dao.UserRepository;
-import com.epam.ta.reportportal.entity.enums.ProjectAttributeEnum;
 import com.epam.ta.reportportal.entity.integration.Integration;
-import com.epam.ta.reportportal.entity.item.issue.IssueType;
-import com.epam.ta.reportportal.entity.project.Project;
-import com.epam.ta.reportportal.entity.project.ProjectIssueType;
-import com.epam.ta.reportportal.entity.user.User;
 import com.epam.ta.reportportal.entity.user.UserType;
 import com.epam.ta.reportportal.model.DeleteBulkRQ;
 import com.epam.ta.reportportal.model.Page;
@@ -57,16 +52,18 @@ import com.epam.ta.reportportal.model.user.RestorePasswordRQ;
 import com.epam.ta.reportportal.ws.BaseMvcTest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Lists;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.http.MediaType;
+import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 /**
  * @author <a href="mailto:ihar_kahadouski@epam.com">Ihar Kahadouski</a>
@@ -86,6 +83,8 @@ class UserControllerTest extends BaseMvcTest {
   @Autowired
   private IssueTypeRepository issueTypeRepository;
 
+  private final String USERS_URL = "/v1/users";
+
   @Test
   void createdUserByIdentityProvider() throws Exception  {
     CreateUserRQFull rq = new CreateUserRQFull();
@@ -99,7 +98,7 @@ class UserControllerTest extends BaseMvcTest {
     rq.setAccountType("SCIM");
 
     MvcResult mvcResult = mockMvc.perform(
-            post("/users").with(token(oAuthHelper.getSuperadminToken()))
+            post("/v1/users").with(token(oAuthHelper.getSuperadminToken()))
                 .contentType(APPLICATION_JSON)
                 .content(objectMapper.writeValueAsBytes(rq))).andExpect(status().isCreated())
         .andReturn();
@@ -135,7 +134,7 @@ class UserControllerTest extends BaseMvcTest {
     rq.setAccountType(UserType.INTERNAL.toString());
 
     MvcResult mvcResult = mockMvc.perform(
-            post("/users").with(token(oAuthHelper.getSuperadminToken()))
+            post(USERS_URL).with(token(oAuthHelper.getSuperadminToken()))
                 .contentType(APPLICATION_JSON)
                 .content(objectMapper.writeValueAsBytes(rq))).andExpect(status().isCreated())
         .andReturn();
@@ -185,7 +184,7 @@ class UserControllerTest extends BaseMvcTest {
     doNothing().when(emailService).sendCreateUserConfirmationEmail(any(), any(), any());
 
     MvcResult mvcResult = mockMvc.perform(
-            post("/users/bid").with(token(oAuthHelper.getDefaultToken()))
+            post(USERS_URL+"/bid").with(token(oAuthHelper.getDefaultToken()))
                 .contentType(APPLICATION_JSON)
                 .content(objectMapper.writeValueAsBytes(rq))).andExpect(status().isCreated())
         .andReturn();
@@ -207,8 +206,9 @@ class UserControllerTest extends BaseMvcTest {
     rq.setFullName("Test User");
     rq.setEmail("test@domain.com");
     MvcResult mvcResult = mockMvc.perform(
-            post("/users/registration?uuid=e5f98deb-8966-4b2d-ba2f-35bc69d30c06").contentType(
-                APPLICATION_JSON).content(objectMapper.writeValueAsBytes(rq)))
+            post(USERS_URL + "/registration?uuid=e5f98deb-8966-4b2d-ba2f-35bc69d30c06")
+                .contentType(APPLICATION_JSON)
+                .content(objectMapper.writeValueAsBytes(rq)))
         .andExpect(status().isCreated()).andReturn();
 
     CreateUserRS createUserRS = objectMapper.readValue(mvcResult.getResponse().getContentAsString(),
@@ -221,20 +221,20 @@ class UserControllerTest extends BaseMvcTest {
 
   @Test
   void getUserBidInfoPositive() throws Exception {
-    mockMvc.perform(get("/users/registration?uuid=e5f98deb-8966-4b2d-ba2f-35bc69d30c06"))
+    mockMvc.perform(get(USERS_URL + "/registration?uuid=e5f98deb-8966-4b2d-ba2f-35bc69d30c06"))
         .andExpect(status().isOk());
   }
 
   @Test
   void deleteUserNegative() throws Exception {
     /* Administrator cannot remove him/her-self */
-    mockMvc.perform(delete("/users/1").with(token(oAuthHelper.getSuperadminToken())))
+    mockMvc.perform(delete(USERS_URL + "/1").with(token(oAuthHelper.getSuperadminToken())))
         .andExpect(status().isForbidden());
   }
 
   @Test
   void deleteUserPositive() throws Exception {
-    mockMvc.perform(delete("/users/2").with(token(oAuthHelper.getSuperadminToken())))
+    mockMvc.perform(delete(USERS_URL + "/2").with(token(oAuthHelper.getSuperadminToken())))
         .andExpect(status().isOk());
   }
 
@@ -244,7 +244,7 @@ class UserControllerTest extends BaseMvcTest {
     DeleteBulkRQ deleteBulkRQ = new DeleteBulkRQ();
     deleteBulkRQ.setIds(Lists.newArrayList(2L));
 
-    mockMvc.perform(delete("/users").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(delete(USERS_URL).with(token(oAuthHelper.getSuperadminToken()))
             .contentType(APPLICATION_JSON)
             .param("ids", "1", "2"))
         .andExpect(status().isOk());
@@ -256,7 +256,7 @@ class UserControllerTest extends BaseMvcTest {
     rq.setFullName("Vasya Pupkin");
     rq.setEmail("defaultemail@domain.com");
     rq.setRole("USER");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(rq))).andExpect(status().isOk());
   }
@@ -266,7 +266,7 @@ class UserControllerTest extends BaseMvcTest {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setEmail("defaltemail@domain.com");
     editUserRQ.setFullName("1");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isBadRequest());
   }
@@ -276,7 +276,7 @@ class UserControllerTest extends BaseMvcTest {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setEmail("defaltemail@domain.com");
     editUserRQ.setFullName(RandomStringUtils.randomAlphabetic(257));
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isBadRequest());
   }
@@ -286,7 +286,7 @@ class UserControllerTest extends BaseMvcTest {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setFullName("Vasya Pupkin");
     editUserRQ.setEmail("superadminemail@domain.com");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().is(409));
   }
@@ -296,7 +296,7 @@ class UserControllerTest extends BaseMvcTest {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setFullName("Vasya Pupkin");
     editUserRQ.setEmail("user1uniquemail@epam.com");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isOk());
   }
@@ -305,7 +305,7 @@ class UserControllerTest extends BaseMvcTest {
   void editAccountTypeByAdmin() throws Exception {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setAccountType("INTERNAL");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isOk());
   }
@@ -314,7 +314,7 @@ class UserControllerTest extends BaseMvcTest {
   void editAccountTypeByAdminNegative() throws Exception {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setAccountType("GITHUB");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isBadRequest());
   }
@@ -323,7 +323,7 @@ class UserControllerTest extends BaseMvcTest {
   void editActiveByAdmin() throws Exception {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setActive(false);
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getSuperadminToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isOk());
   }
@@ -332,7 +332,7 @@ class UserControllerTest extends BaseMvcTest {
   void editAccountTypeByNotAdmin() throws Exception {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setAccountType("INTERNAL");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getDefaultToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isForbidden());
   }
@@ -341,7 +341,7 @@ class UserControllerTest extends BaseMvcTest {
   void editActiveByNotAdmin() throws Exception {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setActive(false);
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(put(USERS_URL + "/default").with(token(oAuthHelper.getDefaultToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isForbidden());
   }
@@ -350,33 +350,33 @@ class UserControllerTest extends BaseMvcTest {
   void editExternalIdByNotAdmin() throws Exception {
     EditUserRQ editUserRQ = new EditUserRQ();
     editUserRQ.setExternalId("test");
-    mockMvc.perform(put("/users/default").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(put(USERS_URL +  "/default").with(token(oAuthHelper.getDefaultToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(editUserRQ))).andExpect(status().isOk());
   }
 
   @Test
   void getUserPositive() throws Exception {
-    mockMvc.perform(get("/users/default").with(token(oAuthHelper.getSuperadminToken())))
+    mockMvc.perform(get(USERS_URL + "/default").with(token(oAuthHelper.getSuperadminToken())))
         .andExpect(status().isOk());
   }
 
   @Test
   void getUserPositiveUsingApiToken() throws Exception {
-    mockMvc.perform(get("/users/default").with(
+    mockMvc.perform(get(USERS_URL + "/default").with(
             token("test__ET4Byc1QUqO8VV8kiCGSP3O4SERb5MJWIowQQ3SiEqHO6hjicoPw-vm1tnrQI5V")))
         .andExpect(status().isOk());
   }
 
   @Test
   void getUsersPositive() throws Exception {
-    mockMvc.perform(get("/users/all").with(token(oAuthHelper.getSuperadminToken())))
+    mockMvc.perform(get(USERS_URL + "/all").with(token(oAuthHelper.getSuperadminToken())))
         .andExpect(status().isOk());
   }
 
   @Test
   void validateUserInfoUsernamePositive() throws Exception {
-    mockMvc.perform(get("/users/registration/info?username=default").with(
+    mockMvc.perform(get(USERS_URL + "/registration/info?username=default").with(
             token(oAuthHelper.getSuperadminToken())))
         .andExpect(status().isOk());
   }
@@ -386,7 +386,7 @@ class UserControllerTest extends BaseMvcTest {
     ChangePasswordRQ rq = new ChangePasswordRQ();
     rq.setOldPassword("password");
     rq.setNewPassword("12345");
-    mockMvc.perform(post("/users/password/change").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(post(USERS_URL + "/password/change").with(token(oAuthHelper.getDefaultToken()))
         .content(objectMapper.writeValueAsBytes(rq))
         .contentType(APPLICATION_JSON)).andExpect(status().isBadRequest());
   }
@@ -400,7 +400,7 @@ class UserControllerTest extends BaseMvcTest {
     when(mailServiceFactory.getDefaultEmailService(true)).thenReturn(emailService);
     doNothing().when(emailService).sendChangePasswordConfirmation(any(), any(), any());
 
-    mockMvc.perform(post("/users/password/change").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(post(USERS_URL + "/password/change").with(token(oAuthHelper.getDefaultToken()))
         .content(objectMapper.writeValueAsBytes(rq))
         .contentType(APPLICATION_JSON)).andExpect(status().isOk());
   }
@@ -411,7 +411,7 @@ class UserControllerTest extends BaseMvcTest {
     rq.setOldPassword("1q2w3e");
     rq.setNewPassword(
         RandomStringUtils.randomAlphabetic(ValidationConstraints.MAX_PASSWORD_LENGTH + 1));
-    mockMvc.perform(post("/users/password/change").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(post(USERS_URL + "/password/change").with(token(oAuthHelper.getDefaultToken()))
         .content(objectMapper.writeValueAsBytes(rq))
         .contentType(APPLICATION_JSON)).andExpect(status().isBadRequest());
   }
@@ -422,7 +422,7 @@ class UserControllerTest extends BaseMvcTest {
     rq.setOldPassword("1q2w3e");
     rq.setNewPassword(
         RandomStringUtils.randomAlphabetic(ValidationConstraints.MIN_PASSWORD_LENGTH - 1));
-    mockMvc.perform(post("/users/password/change").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(post(USERS_URL + "/password/change").with(token(oAuthHelper.getDefaultToken()))
         .content(objectMapper.writeValueAsBytes(rq))
         .contentType(APPLICATION_JSON)).andExpect(status().isBadRequest());
   }
@@ -435,7 +435,7 @@ class UserControllerTest extends BaseMvcTest {
     when(mailServiceFactory.getDefaultEmailService(true)).thenReturn(emailService);
     doNothing().when(emailService).sendRestorePasswordEmail(any(), any(), any(), any());
 
-    mockMvc.perform(post("/users/password/restore").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(post(USERS_URL + "/password/restore").with(token(oAuthHelper.getDefaultToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(restorePasswordRQ))).andExpect(status().isOk());
   }
@@ -445,34 +445,34 @@ class UserControllerTest extends BaseMvcTest {
     final ResetPasswordRQ resetPasswordRQ = new ResetPasswordRQ();
     resetPasswordRQ.setPassword("password");
     resetPasswordRQ.setUuid("e5f98deb-8966-4b2d-ba2f-35bc69d30c06");
-    mockMvc.perform(post("/users/password/reset").with(token(oAuthHelper.getDefaultToken()))
+    mockMvc.perform(post(USERS_URL + "/password/reset").with(token(oAuthHelper.getDefaultToken()))
         .contentType(APPLICATION_JSON)
         .content(objectMapper.writeValueAsBytes(resetPasswordRQ))).andExpect(status().isOk());
   }
 
   @Test
   void isRestorePasswordBidExist() throws Exception {
-    mockMvc.perform(get("/users/password/reset/e5f98deb-8966-4b2d-ba2f-35bc69d30c06").with(
+    mockMvc.perform(get(USERS_URL + "/password/reset/e5f98deb-8966-4b2d-ba2f-35bc69d30c06").with(
             token(oAuthHelper.getDefaultToken()))
         .contentType(APPLICATION_JSON)).andExpect(status().isOk());
   }
 
   @Test
   void getUserProjects() throws Exception {
-    mockMvc.perform(get("/users/default/projects").with(token(oAuthHelper.getDefaultToken())))
+    mockMvc.perform(get(USERS_URL + "/default/projects").with(token(oAuthHelper.getDefaultToken())))
         .andExpect(status().isOk());
   }
 
   @Test
   void getMyself() throws Exception {
-    mockMvc.perform(get("/users").with(token(oAuthHelper.getDefaultToken())))
+    mockMvc.perform(get(USERS_URL).with(token(oAuthHelper.getDefaultToken())))
         .andExpect(status().isOk());
   }
 
   @Test
   void findUsers() throws Exception {
     MvcResult mvcResult = mockMvc.perform(
-            get("/users/search?term=e").with(token(oAuthHelper.getSuperadminToken())))
+            get(USERS_URL + "/search?term=e").with(token(oAuthHelper.getSuperadminToken())))
         .andExpect(status().isOk())
         .andReturn();
     Page userResources = objectMapper.readValue(mvcResult.getResponse().getContentAsString(),
@@ -484,7 +484,30 @@ class UserControllerTest extends BaseMvcTest {
 
   @Test
   void exportUsers() throws Exception {
-    mockMvc.perform(get("/users/export").with(token(oAuthHelper.getSuperadminToken())))
+    mockMvc.perform(get(USERS_URL + "/export").with(token(oAuthHelper.getSuperadminToken())))
+        .andExpect(status().isOk());
+  }
+
+  @Test
+  void userPhoto() throws Exception {
+    final MockHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.multipart(
+            "/v1/data/photo")
+        .file(new MockMultipartFile("file", "file", "image/png",
+            new ClassPathResource("image/image.png").getInputStream()))
+        .contentType(MediaType.MULTIPART_FORM_DATA);
+
+    mockMvc.perform(requestBuilder.with(token(oAuthHelper.getDefaultToken())))
+        .andExpect(status().isOk());
+
+    mockMvc.perform(get("/users/2/avatar?thumbnail=false")
+            .with(token(oAuthHelper.getDefaultToken())))
+        .andExpect(status().isOk());
+
+    mockMvc.perform(get("/users/2/avatar?thumbnail=true").with(
+            token(oAuthHelper.getDefaultToken())))
+        .andExpect(status().isOk());
+
+    mockMvc.perform(delete("/v1/data/photo").with(token(oAuthHelper.getDefaultToken())))
         .andExpect(status().isOk());
   }
 }
